Project Folder/requirements.txt
#  المكتبات المستخدمة لمشروع OCR
Pillow==9.0.1       # مكتبة لمعالجة الصور
pytesseract==0.3.8  # مكتبة للتفاعل مع Tesseract OCR
PyMuPDF==1.19.6    # مكتبة لقراءة وتحليل ملفات PDF

# اللغات المتدرب عليها
# - اللغة الألمانية (German) باستخدام "ISOCP.traineddata"
# - اللغة الإنجليزية (English) باستخدام "ISOCP1.traineddata"


